% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{rapp_icon}
\alias{rapp_icon}
\title{Provide the full path to the "icon" file within the package}
\usage{
rapp_icon(icon_ref)
}
\arguments{
\item{icon_ref}{icon file name generally found in the \code{icon} column
returned via a call to \code{\link[=rappalyze]{rappalyze()}}.}
}
\description{
Provide the full path to the "icon" file within the package
}
\examples{
rapp_icon("Nginx.svg")
}
