% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rappalyze.r
\name{rappalyze}
\alias{rappalyze}
\title{Identify technologies used on a given site/URL}
\usage{
rappalyze(site, quiet = !interactive())
}
\arguments{
\item{site}{either a URL (which will be fetched with `httr::GET()`) or an `httr` `response` object}

\item{verbose}{display progress messages (some sites take a few seconds to analyze)}
}
\description{
Identify technologies used on a given site/URL
}
\examples{
rappalyze("https://rud.is/b")
rappalyze("https://www.quantcast.com")
rappalyze("https://wappalyzer.com")
}
