% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rappalyzer-package.R
\docType{package}
\name{rappalyzer}
\alias{rappalyzer}
\alias{rappalyzer-package}
\title{Identify Technologies Used On Websites}
\description{
The 'Wappalyzer' project <https://wappalyzer.com/> maintains a database
of features that can be compared and combined to identify what technologies -- servers,
frameworks, analytics tools, etc. -- are being used by a target web site. Methods are
provided that mimic the functionality of 'Wappalyzer' in R.
}
\author{
Bob Rudis (bob@rud.is)
}
