% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild.r
\name{rebuild_apps_inventory}
\alias{rebuild_apps_inventory}
\title{Rebuild apps inventory with a different master file}
\usage{
rebuild_apps_inventory(apps_js_file)
}
\arguments{
\item{apps_js_file}{path to a new application inventory file}
}
\description{
The source \code{apps.json} file in the \href{https://raw.githubusercontent.com/AliasIO/Wappalyzer/master/src/apps.json}{Wappalyzer repo}
changes more frequently than the one that comes with the package. Use
this function to rebuild the in-memory data structures that support
the parsing and technology stack detection functions.
}
\examples{
\dontrun{
wapp_url <- "https://raw.githubusercontent.com/AliasIO/Wappalyzer/master/src/apps.json"
tf <- tempfile()
download.file(wapp_url, tf)
rebuild_apps_inventory(tf)
unlink(tf)
}
}
