% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rappalyze.r
\name{rappalyze}
\alias{rappalyze}
\title{Identify technologies used on a given site/URL}
\usage{
rappalyze(site, quiet = !interactive())
}
\arguments{
\item{site}{either a URL (which will be fetched with \code{httr::GET()}) or an \code{httr} \code{response} object}

\item{quiet}{display progress messages (some sites take a few seconds to analyze)}
}
\value{
nested data frame of technology matches or an empty data frame
}
\description{
Filenames returned in the \code{icon} column can be retrieved via the
\code{\link[=rapp_icon]{rapp_icon()}} helper function.
}
\examples{
rappalyze("https://rud.is/b")
rappalyze("https://www.quantcast.com")
rappalyze("https://wappalyzer.com")
}
