% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rappalyze.r
\name{rappalyze}
\alias{rappalyze}
\title{Identify technologies used on a given site/URL}
\usage{
rappalyze(site, agent = NULL)
}
\arguments{
\item{site}{either a URL (which will be fetched with \code{httr::GET()}) or a "minimum vialbe"
\code{httr} \code{response} object (which is a \code{list} with \code{url}, \code{headers} and raw \code{content})}

\item{agent}{user agent to use. Ideally we'd be honest and say we're a bot. We don't since we
are trying to find out what a site is using tech-wise.}
}
\value{
eventually a data frame. for now, a character vector of tech found or \code{character(0)}
}
\description{
Filenames returned in the \code{icon} column can be retrieved via the
\code{\link[=rapp_icon]{rapp_icon()}} helper function.
}
\examples{
rappalyze("https://rud.is/b")
rappalyze("https://www.quantcast.com")
rappalyze("https://wappalyzer.com")
}
